Week 4 Assignment
So far the player can move around, the player will be able to spawn new turrets later, haven't done this feature.

Also, the player will be able to shoot around.

Currently there is a single zombie in the scene, later the scene will respawn zombies for the player to shoot. and The zombies will be able to move around.

Other things in planning:
Both zombie and turret will have health.

There will be a resource, killing zombie will earn resource, and building turrets will cost it.