Week 6 Assignment
About the reward spawn and collection:

Since the game's design is a little different, the reward should be automatically collected once an enemy is defeated.

For this assignment, I did two temp thing to show that I finished the given requirements:

1: respawning: every time you click B (for build), a new turret will be respawned.

2: for now, every time you collide with the Zombie, a message will appear stating the reward is collected.

In the final assignment,

The Zombie will automatically respawned.
The turret built by the player will automatically shoot the enemy.
Once the enemy died, the reward will be collected automatically. (that's why the collision feature is called "shoot")